AIProject_KE/requirements.txt
customtkinter
opencv-python
pillow
tensorflow
numpy
